// src/model/KK/KKLeagueOpponent.ts
export class KKLeagueOpponent {
    boosters: any;
    can_fight: any;
    country: any='';
    country_text: any='';
    girls_count_per_element: any;
    level: any=0;
    match_history: any;
    match_history_sorting: any = 0
    nickname: any='';
    place: any;
    player: any;
    player_league_points: any;
    power: any;
    rewards: any;
    team: any;
}